Squashed commit of the following:
commit 80124eee5ca6842add41ac1a4f7f3b7273331545
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 13:48:34 2025 -0500

    Fix test.

commit 7f3106f65c36f1de0cc745d96a3ca491aeeaad9c
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 13:44:44 2025 -0500

    Fix Admin::GitHubEventsController.

commit 028928173947e457e121debd3cc307389f60b3fc
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 13:41:42 2025 -0500

    Fix GitHubEventsController.

commit 4a418c0da4d308fda9de3fe705fc0102b18f3615
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 13:38:44 2025 -0500

    Fix general settings.

commit c570d0bf6d512d11e373e746549114297088fb7d
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 12:43:28 2025 -0500

    Fix project secret stuff.

commit 5eb07b28feec19b9b1074a40656bb0e9bcbcfd52
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 12:35:17 2025 -0500

    Fix GitHubAccounts.

commit e210bec632c6e321f2d6160e6a60e94868dc5b31
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 12:29:48 2025 -0500

    Fix build cancellation.

commit ad22df210dee6a12fb499c307458e07e32dfe1bc
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 09:12:15 2025 -0500

    Fix build deletion.

commit 136fcad8b032a6469d2054c78da3b0d5b9eb3d74
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 09:06:19 2025 -0500

    Skip authorization for API controllers.

commit 8af56832c062eea954279dc8d28a9b2cdee6d972
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 09:05:35 2025 -0500

    Fix BillingController auth.

commit 406ef5a7e5d98903c8e6d79ac730f1ac0047f412
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 08:14:17 2025 -0500

    Add user impersonation auth.

commit 91a8fa2da2b1f8830885ea6edf818cb8777e1653
Author: Jason Swett <jason@benfranklinlabs.com>
Date:   Wed Feb 5 07:55:26 2025 -0500

    Add verify_authorized.

+++ app/controllers/github_accounts_controller.rb
@@ -4,17 +4,21 @@ class GitHubAccountsController < ApplicationController
   def index
     if user_signed_in?
       @github_accounts = current_user.github_accounts
+      authorize @github_accounts
     else
+      skip_authorization
       redirect_to new_user_session_path
     end
   end
 
   def show
     @github_account = GitHubAccount.find(params[:id])
+    authorize @github_account
   end
 
   def destroy
     github_account = GitHubAccount.find(params[:id])
+    authorize github_account
 
     ActiveRecord::Base.transaction do
       github_account.octokit_client.delete_installation(github_account.github_installation_id)
